Move `h.api.search.transform` to `h.api.transform`
`h.api.storage` needs to be able to use `prepare` from this module
without importing the rest of `h.api.search` (which in turn imports
`h.api.storage`).
