+++ h/api/test/transform_test.py
@@ -2,10 +2,10 @@
 import mock
 import pytest
 
-from h.api.search import transform
+from h.api import transform
 
 
-@mock.patch('h.api.search.transform.groups')
+@mock.patch('h.api.transform.groups')
 def test_prepare_calls_set_group_if_reply(groups):
     annotation = {'permissions': {'read': []}}
 
@@ -14,7 +14,7 @@ def test_prepare_calls_set_group_if_reply(groups):
     groups.set_group_if_reply.assert_called_once_with(annotation)
 
 
-@mock.patch('h.api.search.transform.groups')
+@mock.patch('h.api.transform.groups')
 def test_prepare_calls_insert_group(groups):
     annotation = {'permissions': {'read': []}}
 
@@ -23,7 +23,7 @@ def test_prepare_calls_insert_group(groups):
     groups.insert_group_if_none.assert_called_once_with(annotation)
 
 
-@mock.patch('h.api.search.transform.groups')
+@mock.patch('h.api.transform.groups')
 def test_prepare_calls_set_permissions(groups):
     annotation = {'permissions': {'read': []}}
 
@@ -32,7 +32,7 @@ def test_prepare_calls_set_permissions(groups):
     groups.set_permissions.assert_called_once_with(annotation)
 
 
-@mock.patch('h.api.search.transform.groups')
+@mock.patch('h.api.transform.groups')
 @pytest.mark.parametrize("ann_in,ann_out", [
     # Preserves the basics
     ({}, {}),
@@ -50,7 +50,7 @@ def test_prepare_noop_when_nothing_to_normalize(_, ann_in, ann_out):
     assert ann_in == ann_out
 
 
-@mock.patch('h.api.search.transform.groups')
+@mock.patch('h.api.transform.groups')
 @pytest.mark.parametrize("ann_in,ann_out", [
     ({"target": [{"source": "giraffe"}]},
      {"target": [{"source": "giraffe", "scope": ["*giraffe*"]}]}),
@@ -63,7 +63,7 @@ def test_prepare_adds_scope_field(_, ann_in, ann_out, uri_normalize):
     assert ann_in == ann_out
 
 
-@mock.patch('h.api.search.transform.groups')
+@mock.patch('h.api.transform.groups')
 @pytest.mark.usefixtures("uri_normalize")
 @pytest.mark.parametrize("ann_in,ann_out", [
     ({"uri": "giraffe"},
@@ -103,7 +103,7 @@ def test_prepare_transforms_old_style_comments(groups, ann_in, ann_out):
     assert ann_in == ann_out
 
 
-@mock.patch('h.api.search.transform.groups')
+@mock.patch('h.api.transform.groups')
 @pytest.mark.parametrize("ann,nipsa", [
     ({"user": "george"}, True),
     ({"user": "georgia"}, False),
